pyproject.toml
[project]
name = "nodc-geography"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Magnus Wenzer", email = "magnus.wenzer@smhi.se"},
]
dependencies = [
    "geopandas>=0.14.3",
    "pyyaml>=6.0.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pyinstaller>=4.5.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mypkg = ["."]
